+++ Libraries/SideScroll/Utilities/CompressionUtils.cs
@@ -20,18 +20,24 @@ public class CompressionUtils
 
 		using CallTimer compressCall = call.Timer("Compressing", new Tag("File", fileToCompress.FullName));
 
-		using FileStream originalFileStream = fileToCompress.OpenRead();
+		long compressedSize;
+		using (FileStream originalFileStream = fileToCompress.OpenRead())
+		using (FileStream compressedFileStream = File.Create(fileToCompress.FullName + ".gz"))
+		{
+			// Dispose before reading the size, GZipStream doesn't write the remaining blocks until then
+			// leaveOpen so disposing it doesn't also close the file we're about to measure
+			using (GZipStream compressionStream = new(compressedFileStream, CompressionMode.Compress, leaveOpen: true))
+			{
+				originalFileStream.CopyTo(compressionStream);
+			}
 
-		using FileStream compressedFileStream = File.Create(fileToCompress.FullName + ".gz");
-
-		using GZipStream compressionStream = new(compressedFileStream, CompressionMode.Compress);
-
-		originalFileStream.CopyTo(compressionStream);
+			compressedSize = compressedFileStream.Length;
+		}
 
 		compressCall.Log.Add("Finished Compressing",
 			new Tag("File", fileToCompress.Name),
 			new Tag("Original Size", fileToCompress.Length),
-			new Tag("Compressed Size", compressedFileStream.Length)
+			new Tag("Compressed Size", compressedSize)
 			);
 	}
 
@@ -66,21 +72,26 @@ public class CompressionUtils
 
 	private static void ExtractGzip(CallTimer decompressCall, FileInfo fileToDecompress)
 	{
-		using FileStream originalFileStream = fileToDecompress.OpenRead();
-
 		string currentFileName = fileToDecompress.FullName;
 		string newFileName = currentFileName.Remove(currentFileName.Length - fileToDecompress.Extension.Length);
 
-		using FileStream decompressedFileStream = File.Create(newFileName);
+		long decompressedSize;
+		using (FileStream originalFileStream = fileToDecompress.OpenRead())
+		using (FileStream decompressedFileStream = File.Create(newFileName))
+		{
+			using (GZipStream decompressionStream = new(originalFileStream, CompressionMode.Decompress, leaveOpen: true))
+			{
+				decompressionStream.CopyTo(decompressedFileStream);
+			}
 
-		using GZipStream decompressionStream = new(originalFileStream, CompressionMode.Decompress);
-
-		decompressionStream.CopyTo(decompressedFileStream);
+			decompressedFileStream.Flush();
+			decompressedSize = decompressedFileStream.Length;
+		}
 
 		decompressCall.Log.Add("Finished Decompressing",
 			new Tag("File", fileToDecompress.Name),
-			new Tag("Original Size", fileToDecompress.Length),
-			new Tag("Compressed Size", decompressedFileStream.Length)
+			new Tag("Compressed Size", fileToDecompress.Length),
+			new Tag("Decompressed Size", decompressedSize)
 			);
 	}
 }